Move Decoder and Validation error out of v2
